// src/enums/system-tabs.ts
export enum SYSTEM_TABS {
  INFO = 'info',
  MESSAGE = 'message',
  DOWNLOAD = 'download',
  DOWNLOAD_RAW = 'Raw Data Download',
  DOWNLOAD_LOG = 'Log File Download',
  DOWNLOAD_LGW_INFO = 'LGW Info Download',
  GRAPH = 'graph',
  DIAGNOSIS1 = 'diagnosis1',
  DIAGNOSIS2 = 'diagnosis2',
  STATISTICS = 'statistics',
}